fix(bond): reverse cd ratio scaling
The amount of tranche tokens minted should scale by the debt:collateral
ratio, not the other way around. This means that when the price of the
underlying asset has gone down, the number of tokens the user receives
should scale up. Similarly, when the price of the underlying asset has
gone up, the number of tokens the user receives should scale down.